Refactored code regarding forum details management
