+++ src/constants/Forum.js
@@ -1,0 +1,4 @@
+export const manageForumDetailsType = {
+  CREATE: 0,
+  EDIT: 1
+};
